added some todos and fixed import
