+++ x-pack/plugins/files/server/file/file_attributes_reducer.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { FileSavedObjectAttributes } from '../../common';
+
+export type Action =
+  | { action: 'create'; payload: FileSavedObjectAttributes }
+  | {
+      action: 'uploading';
+      payload?: undefined;
+    }
+  | { action: 'uploaded'; payload: { content_ref: string; size: number } }
+  | { action: 'uploadError'; payload?: undefined };
+
+export function createDefaultFileAttributes(): Pick<
+  FileSavedObjectAttributes,
+  'created_at' | 'updated_at' | 'status'
+> {
+  const dateString = new Date().toISOString();
+  return {
+    created_at: dateString,
+    updated_at: dateString,
+    status: 'AWAITING_UPLOAD',
+  };
+}
+
+export function fileAttributesReducer(
+  state: FileSavedObjectAttributes,
+  { action, payload }: Action
+): FileSavedObjectAttributes {
+  switch (action) {
+    case 'uploading':
+      return { ...state, content_ref: undefined, status: 'UPLOADING' };
+    case 'uploaded':
+      return { ...state, ...payload, status: 'READY' };
+    case 'uploadError':
+      return { ...state, status: 'ERROR', content_ref: undefined };
+    default:
+      return state;
+  }
+}
