+++ x-pack/plugins/files/server/file/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { File } from './file';
+export {
+  Action,
+  createDefaultFileAttributes,
+  fileAttributesReducer,
+} from './file_attributes_reducer';
