+++ x-pack/plugins/files/server/integration_tests/file_service.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { Readable } from 'stream';
-import { InternalFileService } from './file_service';
+import { InternalFileService } from '../file_service';
 
 describe('FileService', () => {
   let fileService: InternalFileService;
